pyproject.toml
[tool.poetry]
name = "web-api"
version = "0.1.0"
description = ""
authors = ["Daniil Fedyaev <wintercitizen@evrone.com>"]

[tool.poetry.dependencies]
python = "^3.8"
funcy = "^1.14"
uvloop = "^0.14.0"
pydantic = "^1.7.2"
fastapi = "^0.61.2"
uvicorn = "^0.13.0"
motor = "^2.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aioredis = "^1.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"
mypy = "^0.770"
black = "^19.10b0"
ipython = "^7.13.0"
ipdb = "^0.13.2"
factory_boy = "^2.12.0"
rope = "^0.16.0"
syrupy = "^1.0.0"
httpx = "^0.16.1"
pytest-aiohttp = "^0.3.0"
pytest-asyncio = "^0.14.0"

[tool.black]
line-length = 79
py38 = true
skip-string-normalization = true
exclude = '(.git)'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
